Basic Data puller script
Currently it can take a YouTube video link and pull the caption,description and sponsor block data as well as some general information about the video. I plan to try and add some more functionality and error handling so that we can get all of a channels videos, but this is a good start I think, has 1 video downloaded as an example of the current format and files. If any changes are needed let me know.